refactor(cdk-mcp-server): rename genai_cdk_github_loader.py to genai_cdk_loader.py and update imports
